requirements.txt
Django==2.0.4
django-crispy-forms==1.7.2
djangorestframework==3.8.2
Pillow==5.1.0
pytz==2018.4
